// web/src/api/generated/StatusSegment.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeStatus } from "./NodeStatus";

/**
 * A span of constant status in a node's history.
 */
export type StatusSegment = { from_ms: number, to_ms: number, status: NodeStatus, };